#14: Move Props to i.g.d.t.arg.scalar package
