+++ src/main/java/io/github/dgroup/term4j/arg/scalar/Props.java
@@ -22,28 +22,22 @@
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.github.dgroup.term4j.arg;
+package io.github.dgroup.term4j.arg.scalar;
 
-import org.hamcrest.core.IsEqual;
-import org.junit.Test;
-import org.llorllale.cactoos.matchers.Assertion;
+import java.util.Properties;
+import org.cactoos.Scalar;
 
 /**
- * Test case for {@link Props}.
+ * The application properties.
+ *
+ * The properties might be specified using <em>-D</em> option.
  *
  * @since 0.2.0
- * @checkstyle MagicNumberCheck (500 lines)
- * @checkstyle JavadocMethodCheck (500 lines)
  */
-@SuppressWarnings("PMD.AvoidDuplicateLiterals")
-public final class PropsTest {
+public final class Props implements Scalar<Properties> {
 
-    @Test
-    public void value() {
-        new Assertion<>(
-            "fetch the application properties including custom one",
-            () -> new Props().value().getProperty("propof"),
-            new IsEqual<>("10")
-        ).affirm();
+    @Override
+    public Properties value() {
+        return System.getProperties();
     }
 }
